apps/api/src/services/forecast_service.py
"""
Forecast service for StockSense AI.

Provides business logic for demand forecasting operations including
model training, prediction, and accuracy analysis.
"""

from typing import List, Optional, Dict, Any
from datetime import datetime, timedelta
from sqlalchemy.ext.asyncio import AsyncSession
from sqlalchemy import select, func, and_, or_
import structlog
import pandas as pd
import numpy as np
from sklearn.ensemble import RandomForestRegressor
from sklearn.metrics import mean_absolute_percentage_error, mean_absolute_error
import joblib
import os

from ..models.user import User
from ..schemas.forecast import (
    ForecastCreate, ForecastUpdate, ForecastResponse,
    ForecastModel, ForecastAccuracy, ForecastRequest
)

logger = structlog.get_logger()

class ForecastService:
    """Service for demand forecasting operations."""
    
    def __init__(self, db: AsyncSession):
        self.db = db
        self.models_dir = "models"
        os.makedirs(self.models_dir, exist_ok=True)
    
    async def create_forecast(self, forecast_data: ForecastCreate, user: User) -> ForecastResponse:
        """Create a new forecast."""
        try:
            # For now, we'll create a placeholder forecast
            # In a real implementation, this would trigger model training
            forecast = {
                "id": "fc_" + str(int(datetime.utcnow().timestamp())),
                "item_id": forecast_data.item_id,
                "location_id": forecast_data.location_id,
                "forecast_period": forecast_data.forecast_period,
                "forecast_horizon": forecast_data.forecast_horizon,
                "model_type": forecast_data.model_type,
                "forecast_values": self._generate_sample_forecast(forecast_data.forecast_horizon),
                "confidence_intervals": self._generate_confidence_intervals(forecast_data.forecast_horizon),
                "accuracy_metrics": {
                    "mape": 0.15,
                    "mae": 10.5,
                    "rmse": 12.3
                },
                "status": "completed",
                "created_by": str(user.id),
                "created_at": datetime.utcnow()
            }
            
            logger.info("Forecast created", forecast_id=forecast["id"])
            return ForecastResponse(**forecast)
            
        except Exception as e:
            logger.error("Failed to create forecast", error=str(e))
            raise
    
    async def get_forecasts(
        self, 
        user: User,
        item_id: Optional[str] = None,
        location_id: Optional[str] = None,
        skip: int = 0,
        limit: int = 100
    ) -> List[ForecastResponse]:
        """Get forecasts with filtering and pagination."""
        try:
            # In a real implementation, this would query the database
            # For now, return sample data
            forecasts = []
            for i in range(min(limit, 5)):
                forecast = {
                    "id": f"fc_{i+1}",
                    "item_id": f"item_{i+1}",
                    "location_id": f"loc_{i+1}",
                    "forecast_period": "daily",
                    "forecast_horizon": 30,
                    "model_type": "random_forest",
                    "forecast_values": self._generate_sample_forecast(30),
                    "confidence_intervals": self._generate_confidence_intervals(30),
                    "accuracy_metrics": {
                        "mape": round(0.1 + i * 0.02, 3),
                        "mae": round(8.0 + i * 1.5, 1),
                        "rmse": round(10.0 + i * 2.0, 1)
                    },
                    "status": "completed",
                    "created_by": str(user.id),
                    "created_at": datetime.utcnow() - timedelta(days=i)
                }
                forecasts.append(ForecastResponse(**forecast))
            
            return forecasts
            
        except Exception as e:
            logger.error("Failed to get forecasts", error=str(e))
            raise
    
    async def get_forecast_accuracy(self, user: User, days: int = 30) -> ForecastAccuracy:
        """Get forecast accuracy metrics."""
        try:
            # In a real implementation, this would calculate actual accuracy
            # For now, return sample metrics
            accuracy = ForecastAccuracy(
                overall_mape=0.125,
                overall_mae=9.8,
                overall_rmse=11.2,
                item_level_accuracy=[
                    {
                        "item_id": "item_1",
                        "item_name": "Sample Item 1",
                        "mape": 0.12,
                        "mae": 8.5,
                        "rmse": 10.1
                    },
                    {
                        "item_id": "item_2", 
                        "item_name": "Sample Item 2",
                        "mape": 0.15,
                        "mae": 11.2,
                        "rmse": 13.8
                    }
                ],
                model_performance=[
                    {
                        "model_type": "random_forest",
                        "mape": 0.13,
                        "mae": 9.1,
                        "rmse": 11.5
                    },
                    {
                        "model_type": "arima",
                        "mape": 0.14,
                        "mae": 10.2,
                        "rmse": 12.8
                    }
                ]
            )
            
            return accuracy
            
        except Exception as e:
            logger.error("Failed to get forecast accuracy", error=str(e))
            raise
    
    async def train_model(self, model_data: Dict[str, Any], user: User) -> str:
        """Train a new forecasting model."""
        try:
            # In a real implementation, this would:
            # 1. Load historical data
            # 2. Preprocess the data
            # 3. Train the model
            # 4. Save the model
            # 5. Return model ID
            
            model_id = f"model_{int(datetime.utcnow().timestamp())}"
            
            # Simulate model training
            logger.info("Training model", model_id=model_id, model_type=model_data.get("model_type"))
            
            # Save model metadata
            model_metadata = {
                "model_id": model_id,
                "model_type": model_data.get("model_type", "random_forest"),
                "training_date": datetime.utcnow().isoformat(),
                "parameters": model_data.get("parameters", {}),
                "created_by": str(user.id)
            }
            
            # In a real implementation, save the actual trained model
            # joblib.dump(model, os.path.join(self.models_dir, f"{model_id}.pkl"))
            
            logger.info("Model training completed", model_id=model_id)
            return model_id
            
        except Exception as e:
            logger.error("Failed to train model", error=str(e))
            raise
    
    async def predict_demand(self, forecast_request: ForecastRequest, user: User) -> Dict[str, Any]:
        """Generate demand predictions."""
        try:
            # In a real implementation, this would:
            # 1. Load the trained model
            # 2. Preprocess input data
            # 3. Generate predictions
            # 4. Return results
            
            # Simulate prediction
            horizon = forecast_request.forecast_horizon
            base_demand = forecast_request.base_demand or 100
            
            # Generate realistic demand pattern
            np.random.seed(42)  # For reproducible results
            trend = np.linspace(0, 0.1, horizon)  # Slight upward trend
            seasonality = 0.1 * np.sin(2 * np.pi * np.arange(horizon) / 7)  # Weekly seasonality
            noise = np.random.normal(0, 0.05, horizon)  # Random noise
            
            predictions = base_demand * (1 + trend + seasonality + noise)
            predictions = np.maximum(predictions, 0)  # Ensure non-negative
            
            # Generate confidence intervals
            confidence_level = 0.95
            std_dev = np.std(predictions) * 0.1
            z_score = 1.96  # 95% confidence interval
            
            lower_bound = predictions - z_score * std_dev
            upper_bound = predictions + z_score * std_dev
            
            result = {
                "predictions": predictions.tolist(),
                "confidence_intervals": {
                    "lower": lower_bound.tolist(),
                    "upper": upper_bound.tolist(),
                    "confidence_level": confidence_level
                },
                "model_info": {
                    "model_type": forecast_request.model_type,
                    "training_date": datetime.utcnow().isoformat(),
                    "features_used": ["historical_demand", "seasonality", "trend"]
                },
                "forecast_horizon": horizon,
                "generated_at": datetime.utcnow().isoformat()
            }
            
            logger.info("Demand prediction generated", 
                       item_id=forecast_request.item_id,
                       horizon=horizon)
            
            return result
            
        except Exception as e:
            logger.error("Failed to generate demand prediction", error=str(e))
            raise
    
    def _generate_sample_forecast(self, horizon: int) -> List[float]:
        """Generate sample forecast values."""
        np.random.seed(42)
        base_value = 100
        trend = np.linspace(0, 0.2, horizon)
        seasonality = 0.1 * np.sin(2 * np.pi * np.arange(horizon) / 7)
        noise = np.random.normal(0, 0.05, horizon)
        
        forecast = base_value * (1 + trend + seasonality + noise)
        return np.maximum(forecast, 0).tolist()
    
    def _generate_confidence_intervals(self, horizon: int) -> Dict[str, List[float]]:
        """Generate sample confidence intervals."""
        forecast = self._generate_sample_forecast(horizon)
        std_dev = np.std(forecast) * 0.1
        
        lower = [max(0, f - 1.96 * std_dev) for f in forecast]
        upper = [f + 1.96 * std_dev for f in forecast]
        
        return {
            "lower": lower,
            "upper": upper,
            "confidence_level": 0.95
        }
